\name{   agjack.pi0       }
\alias{    agjack.pi0     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
 averaged generalized jackknife estimate of pi0
}
\description{
 averaged generalized jackknife estimate of pi0

}
\usage{
agjack.pi0(subtobj,mean.n=mean,pointpair=FALSE, trunc=TRUE,tol=1e-5)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{subtobj}{              }
  \item{mean.n}{              }
  \item{pointpair}{              }
  \item{trunc}{              }
  \item{tol}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
