\name{   geomean       }
\alias{    geomean     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
geometric mean

}
\description{
geometric mean

}
\usage{
geomean(x)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{numeric vector}
%  \item{}{              }
%  \item{}{              }
}
%\details{
%
%
%
%
%
%}
\value{numeric scalar
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
%\references{
%
%
%}
\author{ Long Qu [\email{longor@iastate.edu}] }
%\note{
%
%
%}
%\seealso{ 
%
%
%}
%\examples{
%\dontrun{
%
%
%}
%
%
%
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
