\name{   gjack       }
\alias{    gjack     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
 generalized jackknife
}
\description{
 generalized jackknife

}
\usage{
gjack(theta1, theta2, R)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theta1}{Numeric              }
  \item{theta2}{     Numeric         }
  \item{R}{         Numeric     }
}
\details{
Computes \code{(theta1-R*theta2)/(1-R)}




}
\value{Numeric
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
%\references{
%
%
%}
\author{ Long Qu [\email{longor@iastate.edu}] }
%\note{
%
%
%}
%\seealso{ 
%
%
%}
%\examples{
%\dontrun{
%
%
%}
%
%
%
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
