\name{   harmean       }
\alias{    harmean     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
harmonic mean

}
\description{
harmonic mean

}
\usage{
harmean(x)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{              }
%  \item{}{              }
%  \item{}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
