\name{   histf1       }
\alias{    histf1     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
 histf1
}
\description{
histf1
}
\usage{
histf1(p,max.bins=20,bin.method=c("max","nmse","bootstrap","Sturges","Scott","FD"),
        discrete=FALSE,seq.perm=FALSE,nboots=200,plotit=FALSE,perm.n,perm.h,...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p}{              }
  \item{max.bins}{              }
  \item{bin.method}{              }
  \item{discrete}{              }
  \item{seq.perm}{              }
  \item{nboots}{              }
  \item{plotit}{              }
  \item{perm.n}{              }
  \item{perm.h}{              }
  \item{...}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
