\name{   load.cdf       }
\alias{    load.cdf     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 

loading pre-computed conditional cdf objects from memory or disk
}
\description{
loading pre-computed conditional cdf objects from memory or disk

}
\usage{
load.cdf(path=getwd())
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  
  \item{path}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
