\name{   NBsplines       }
\alias{    NBsplines     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 

normalized B-splines
}
\description{
normalized B-splines
}
\usage{
NBsplines(x,knots,degree)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  
  \item{x}{              }
  \item{knots}{              }
  \item{degree}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
