\name{   ncpest.iter       }
\alias{    ncpest.iter     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
iterated ncp estimates
}
\description{
iterated ncp estimates

}
\usage{
ncpest.iter(p,estimates=c("all","compromise","pi0","f1"),iter=2,weights,
           eps=1e-6,keep.cdf=TRUE,...)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p}{              }
  \item{estimates}{              }
  \item{iter}{              }
  \item{weights}{              }
  \item{eps}{              }
  \item{keep.cdf}{              }
  \item{\dots}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
