\name{   pavaf1       }
\alias{    pavaf1     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
 pavaf1
}
\description{
pavaf1
}
\usage{
pavaf1(p,max.bins=20,bin.method=c("max","Sturges","Scott","FD"),
                discrete=FALSE,plotit=FALSE,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p}{              }
  \item{max.bins}{              }
  \item{bin.method}{              }
  \item{discrete}{              }
  \item{plotit}{              }
  \item{...}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
