\name{internal functions not to be used by users directly}
\alias{combnBt2Rc}
\alias{combnBtRc}
\alias{combnRRepl}
\alias{combn2RRepl}
\alias{combnBt2RAllDifc}
\alias{momeff.zeromean}
\alias{momeff.nonzeromean}
\alias{momeff.nonzeromean}
\alias{extrp.pi0.only}
\alias{extrp.pi0.slope}
\alias{extrp.pi0.rate}
\alias{extrp.pi0.both}
\alias{extrp.pi0.gam2}
\alias{extrp.pi0.slope.gam2}
\alias{extrp.pi0.rate.gam2}
\alias{extrp.pi0.both.gam2}

\title{internal functions not to be used by users directly}
\description{
  Internal Functions Not to Be Used by Users Directly.
}
\details{
  These are not to be called by the user. . 
}
\keyword{ internal }
