\name{   varB       }
\alias{    varB     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
Variance of the reflected Bsplines

}
\description{

2nd order raw moment of the Bsplines
}
\usage{
varB(m,deltamax,K)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{m}{              }
  \item{deltamax}{              }
  \item{K}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
