\name{   cond.cdf       }
\alias{    cond.cdf     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
conditional cdf of p-values given ncp

}
\description{
conditional cdf of p-values given ncp

}
\usage{
cond.cdf(p.eval,ncp,test=c("t","z"),alternative=c("two.sided","less","greater"),
    df=if(test=="z")Inf else df,keep.cdf=TRUE,suppressWarnings=TRUE) #,save=FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p.eval}{              }
  \item{ncp}{              }
  \item{test}{              }
  \item{alternative}{              }
  \item{df}{              }
  \item{keep.cdf}{              }
  \item{suppressWarnings}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
