\name{   gjack       }
\alias{    gjack     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
 generalized jackknife
}
\description{
 generalized jackknife

}
\usage{
gjack(theta1, theta2, R)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theta1}{              }
  \item{theta2}{              }
  \item{R}{              }
}
\details{





}
\value{
%  \item{}{              }
%  \item{}{              }
%  \item{}{              }
  
}
\references{


}
\author{ Long Qu [\email{longor@iastate.edu}] }
\note{


}
\seealso{ 


}
\examples{
\dontrun{


}



}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar  }
%\keyword{  }
%\keyword{  }
