\name{internal functions not to be used by users directly}
\alias{combnBt2Rc}
\alias{combnBtRc}
\alias{combnRRepl}
\alias{combn2RRepl}
\alias{combnBt2RAllDifc}
\alias{momeff.zeromean}
\alias{momeff.nonzeromean}
\alias{momeff.nonzeromean}
\alias{NBsplines}
\alias{dFsnc.mix}
\alias{dncp.nparncpF}
\alias{dncp.parncpF}
\alias{dncp.sparncpF}
\alias{fitted.nparncpF}
\alias{fitted.parncpF}
\alias{fitted.sparncpF}
\alias{fitted.values.nparncpF}
\alias{fitted.values.parncpF}
\alias{fitted.values.sparncpF}
\alias{nparncpF}
\alias{nparncpF.sqp}
\alias{parncpF}
\alias{parncpF.lbfgsb.central}
\alias{parncpF.lbfgsb.noncentral}
\alias{parncpF.unconstrained.central}
\alias{parncpF.unconstrained.noncentral}
\alias{plot.nparncpF}
\alias{plot.parncpF}
\alias{plot.sparncpF}
\alias{print.nparncpF}
\alias{print.parncpF}
\alias{print.sparncpF}
\alias{sparncpF}
\alias{sparncpF.nparncpF}
\alias{sparncpF.numeric}
\alias{sparncpF.parncpF}
\alias{summary.nparncpF}
\alias{summary.parncpF}
\alias{summary.sparncpF}

\alias{lfdr2p}

\title{internal functions not to be used by users directly}
\description{
  Internal Functions Not to Be Used by Users Directly.
}
\details{
  These are not to be called by the user. . 
}
\keyword{ internal }
