\name{   varB       }
\alias{    varB     }
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ 
Variance of the reflected Bsplines

}
\description{

2nd order raw moment of the Bsplines
}
\usage{
varB(m,deltamax,K)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{m}{The \code{m}th spline function to which variance is to be computed}
  \item{deltamax}{maximum noncentrality parameter}
  \item{K}{Number of splines}
}
%\details{
%}
\value{Numeric scalar
}
%\references{
%
%
%}
\author{ Long Qu }
%\note{
%
%
%}
\seealso{ \code{\link{NBsplines}}


}
%\examples{
%\dontrun{
%
%
%}
%
%
%
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{smooth  }
%\keyword{  }
%\keyword{  }
